+++ src/Data/PEM/Types.hs
@@ -1,3 +1,6 @@
+{-# LANGUAGE DeriveAnyClass #-}
+{-# LANGUAGE DeriveGeneric  #-}
+
 {- |
 Module      : Data.PEM.Types
 License     : BSD-style
@@ -11,7 +14,9 @@ module Data.PEM.Types
   ) where
 
 import           Basement.NormalForm ( NormalForm (..) )
+import           Control.DeepSeq ( NFData (..) )
 import           Data.ByteString ( ByteString )
+import           GHC.Generics ( Generic )
 
 -- | A type representing single PEM sections.
 data PEM = PEM
@@ -23,8 +28,11 @@ data PEM = PEM
   , pemContent :: ByteString
     -- ^ Binary content of the section.
   }
-  deriving (Eq, Show)
+  deriving (Eq, Generic, NFData, Show)
 
+-- | The t'PEM' instance of 'NormalForm' is deprecated and will be removed from
+-- a future version of this package. See the 'NFData' type class exported by
+-- "Control.DeepSeq" of the @deepseq@ GHC boot package.
 instance NormalForm PEM where
   toNormalForm pem =
     toNormalForm (pemName pem) `seq` nfLbs (pemHeader pem) `seq` pemContent pem `seq` ()